requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.2
rich==13.7.0 